+++ rlberry/agents/torch/utils/models.py
@@ -194,6 +194,16 @@ class BaseModule(torch.nn.Module):
 
 
 class Table(torch.nn.Module):
+    """Torch module for a policy for discrete state-action spaces.
+
+    Parameters
+    ----------
+    state_size: int
+        Number of states
+    action_size: int
+        Number of actions
+    """
+
     def __init__(self, state_size, action_size):
         super().__init__()
         self.policy = nn.Embedding.from_pretrained(
@@ -210,6 +220,26 @@ class Table(torch.nn.Module):
 
 
 class MultiLayerPerceptron(BaseModule):
+    """Torch module for an MLP.
+
+    Parameters
+    ----------
+    in_size: int
+        Input size
+    layer_sizes: Sequence[int]
+        Dimensions of each hidden layer.
+    reshape: bool, default = True
+        If True, input tensors are reshaped to (batch_size, dim)
+    out_size: int, optional
+        Output size. If None, the output size is given by the last
+        element of layer_sizes.
+    activation: {"RELU", "TANH", "ELU"}
+        Activation function.
+    is_policy: bool, default=False
+        If true, the :meth:`forward` method returns a categorical
+        distribution corresponding to the softmax of the output.
+    """
+
     def __init__(
         self,
         in_size=None,
@@ -223,6 +253,7 @@ class MultiLayerPerceptron(BaseModule):
         super().__init__(**kwargs)
         self.reshape = reshape
         self.layer_sizes = layer_sizes or [64, 64]
+        self.layer_sizes = list(self.layer_sizes)
         self.out_size = out_size
         self.activation = activation_factory(activation)
         self.is_policy = is_policy
@@ -258,6 +289,25 @@ class MultiLayerPerceptron(BaseModule):
 
 
 class DuelingNetwork(BaseModule):
+    """Torch module for a DQN dueling network based on a MultiLayerPerceptron.
+
+    Parameters
+    -----------
+    in_size: int
+        Input size
+    base_module_kwargs: dict
+        Parameters for :func:`~rlberry.agents.torch.utils.training.model_factory`
+        to build shared (MLP) architecture for the advantage and value nets.
+    value_kwargs: dict
+        Parameters for :func:`~rlberry.agents.torch.utils.training.model_factory`
+        to build value network (MLP).
+    advantage_kwargs: dict
+        Parameters for :func:`~rlberry.agents.torch.utils.training.model_factory`
+        to build advantage network (MLP).
+    out_size: int
+        Output size.
+    """
+
     def __init__(
         self,
         in_size=None,
@@ -290,6 +340,32 @@ class DuelingNetwork(BaseModule):
 
 
 class ConvolutionalNetwork(nn.Module):
+    """Torch module for a CNN.
+
+    Expects inputs of shape BCHW, where
+    B = batch size;
+    C = number of channels;
+    H = height;
+    W = width.
+
+    Parameters
+    ----------
+    activation: {"RELU", "TANH", "ELU"}
+        Activation function.
+    in_channels: int
+        Number of input channels C
+    in_height: int
+        Input height H
+    in_width: int
+        Input width W
+    head_mlp_kwargs: dict, optional
+        Parameters to build an MLP
+        (:class:`~rlberry.agents.torch.utils.models.MultiLayerPerceptron`)
+        using the factory
+        :func:`~rlberry.agents.torch.utils.training.model_factory`
+
+    """
+
     def __init__(
         self,
         activation="RELU",
@@ -339,8 +415,12 @@ class ConvolutionalNetwork(nn.Module):
 
     def forward(self, x):
         """
-            Forward convolutional network
-        :param x: tensor of shape BCHW
+        Forward convolutional network
+
+        Parameters
+        ----------
+        x: torch.tensor
+            Tensor of shape BCHW
         """
         return self.head(self.convolutions(x))